feat: 닉네임 검증 status 파일로 추출
